+++ test/controllers/amends/AmendHumanMedicinePlasticPackagingControllerSpec.scala
@@ -14,49 +14,50 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.amends
 
 import base.SpecBase
 import connectors.CacheConnector
-import forms.AmendImportedPlasticPackagingFormProvider
+import forms.AmendHumanMedicinePlasticPackagingFormProvider
 import models.NormalMode
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.amends.AmendImportedPlasticPackagingPage
+import pages.amends.AmendHumanMedicinePlasticPackagingPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import views.html.AmendImportedPlasticPackagingView
+import views.html.amends.AmendHumanMedicinePlasticPackagingView
 
 import scala.concurrent.Future
 
-class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
+class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
 
-  val formProvider = new AmendImportedPlasticPackagingFormProvider()
+  val formProvider = new AmendHumanMedicinePlasticPackagingFormProvider()
   val form         = formProvider()
 
   def onwardRoute = Call("GET", "/foo")
 
   val validAnswer = 0
 
-  lazy val amendImportedPlasticPackagingRoute =
-    routes.AmendImportedPlasticPackagingController.onPageLoad(NormalMode).url
 
-  "AmendImportedPlasticPackaging Controller" - {
+  lazy val amendHumanMedicinePlasticPackagingRoute =
+    routes.AmendHumanMedicinePlasticPackagingController.onPageLoad(NormalMode).url
+
+  "AmendHumanMedicinePlasticPackaging Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
       val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, NormalMode, taxReturnOb)(request,
@@ -67,14 +68,15 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val ans = userAnswers.set(AmendImportedPlasticPackagingPage, validAnswer).success.value
+      val ans = userAnswers.
+        set(AmendHumanMedicinePlasticPackagingPage, validAnswer).success.value
 
       val application = applicationBuilder(userAnswers = Some(ans)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
 
-        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
 
         val result = route(application, request).value
 
@@ -90,7 +92,7 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
 
       val mockCacheConnector = mock[CacheConnector]
 
-      when(mockCacheConnector.set(any() ,any())(any())) thenReturn Future.successful(mockResponse)
+      when(mockCacheConnector.set(any(), any())(any())) thenReturn Future.successful(mockResponse)
 
       val application =
         applicationBuilder(userAnswers = Some(userAnswers))
@@ -101,7 +103,7 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
 
       running(application) {
         val request =
-          FakeRequest(POST, amendImportedPlasticPackagingRoute)
+          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
@@ -116,7 +118,7 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
       val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
 
         val result = route(application, request).value
 
@@ -130,18 +132,18 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
 
       running(application) {
         val request =
-          FakeRequest(POST, amendImportedPlasticPackagingRoute)
+          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", "invalid value"))
 
         val boundForm = form.bind(Map("value" -> "invalid value"))
 
-        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
 
         val result = route(application, request).value
 
         status(result) mustEqual BAD_REQUEST
         contentAsString(result) mustEqual view(boundForm, NormalMode, taxReturnOb)(request,
-                                                                      messages(application)
+          messages(application)
         ).toString
       }
     }
@@ -151,12 +153,12 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
       val application = applicationBuilder(userAnswers = None).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
 
@@ -166,14 +168,14 @@ class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoS
 
       running(application) {
         val request =
-          FakeRequest(POST, amendImportedPlasticPackagingRoute)
+          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
 
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
   }
