+++ test/controllers/amends/AmendImportedPlasticPackagingControllerSpec.scala
@@ -14,49 +14,49 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.amends
 
 import base.SpecBase
 import connectors.CacheConnector
-import forms.AmendManufacturedPlasticPackagingFormProvider
+import forms.AmendImportedPlasticPackagingFormProvider
 import models.NormalMode
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.amends.AmendManufacturedPlasticPackagingPage
+import pages.amends.AmendImportedPlasticPackagingPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import views.html.AmendManufacturedPlasticPackagingView
+import views.html.amends.AmendImportedPlasticPackagingView
 
 import scala.concurrent.Future
 
-class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
+class AmendImportedPlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
 
-  val formProvider = new AmendManufacturedPlasticPackagingFormProvider()
+  val formProvider = new AmendImportedPlasticPackagingFormProvider()
   val form         = formProvider()
 
   def onwardRoute = Call("GET", "/foo")
 
   val validAnswer = 0
 
-  lazy val amendManufacturedPlasticPackagingRoute =
-    routes.AmendManufacturedPlasticPackagingController.onPageLoad(NormalMode).url
+  lazy val amendImportedPlasticPackagingRoute =
+    routes.AmendImportedPlasticPackagingController.onPageLoad(NormalMode).url
 
-  "AmendManufacturedPlasticPackaging Controller" - {
+  "AmendImportedPlasticPackaging Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
       val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendManufacturedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[AmendManufacturedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, NormalMode, taxReturnOb)(request,
@@ -67,14 +67,14 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val ans = userAnswers.set(AmendManufacturedPlasticPackagingPage, validAnswer).success.value
+      val ans = userAnswers.set(AmendImportedPlasticPackagingPage, validAnswer).success.value
 
       val application = applicationBuilder(userAnswers = Some(ans)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendManufacturedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
 
-        val view = application.injector.instanceOf[AmendManufacturedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
 
         val result = route(application, request).value
 
@@ -90,7 +90,7 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
 
       val mockCacheConnector = mock[CacheConnector]
 
-      when(mockCacheConnector.set(any(), any())(any())) thenReturn Future.successful(mockResponse)
+      when(mockCacheConnector.set(any() ,any())(any())) thenReturn Future.successful(mockResponse)
 
       val application =
         applicationBuilder(userAnswers = Some(userAnswers))
@@ -101,7 +101,7 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
 
       running(application) {
         val request =
-          FakeRequest(POST, amendManufacturedPlasticPackagingRoute)
+          FakeRequest(POST, amendImportedPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
@@ -116,7 +116,7 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
       val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendManufacturedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
 
         val result = route(application, request).value
 
@@ -130,18 +130,18 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
 
       running(application) {
         val request =
-          FakeRequest(POST, amendManufacturedPlasticPackagingRoute)
+          FakeRequest(POST, amendImportedPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", "invalid value"))
 
         val boundForm = form.bind(Map("value" -> "invalid value"))
 
-        val view = application.injector.instanceOf[AmendManufacturedPlasticPackagingView]
+        val view = application.injector.instanceOf[AmendImportedPlasticPackagingView]
 
         val result = route(application, request).value
 
         status(result) mustEqual BAD_REQUEST
         contentAsString(result) mustEqual view(boundForm, NormalMode, taxReturnOb)(request,
-          messages(application)
+                                                                      messages(application)
         ).toString
       }
     }
@@ -151,12 +151,12 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
       val application = applicationBuilder(userAnswers = None).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendManufacturedPlasticPackagingRoute)
+        val request = FakeRequest(GET, amendImportedPlasticPackagingRoute)
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
 
@@ -166,14 +166,14 @@ class AmendManufacturedPlasticPackagingControllerSpec extends SpecBase with Mock
 
       running(application) {
         val request =
-          FakeRequest(POST, amendManufacturedPlasticPackagingRoute)
+          FakeRequest(POST, amendImportedPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
 
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
   }
