+++ test/controllers/amends/AmendRecycledPlasticPackagingControllerSpec.scala
@@ -14,50 +14,49 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.amends
 
 import base.SpecBase
 import connectors.CacheConnector
-import forms.AmendHumanMedicinePlasticPackagingFormProvider
+import forms.AmendRecycledPlasticPackagingFormProvider
 import models.NormalMode
 import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.amends.AmendHumanMedicinePlasticPackagingPage
+import pages.amends.AmendRecycledPlasticPackagingPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import views.html.AmendHumanMedicinePlasticPackagingView
+import views.html.amends.AmendRecycledPlasticPackagingView
 
 import scala.concurrent.Future
 
-class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
+class AmendRecycledPlasticPackagingControllerSpec extends SpecBase with MockitoSugar {
 
-  val formProvider = new AmendHumanMedicinePlasticPackagingFormProvider()
+  val formProvider = new AmendRecycledPlasticPackagingFormProvider()
   val form         = formProvider()
 
   def onwardRoute = Call("GET", "/foo")
 
   val validAnswer = 0
 
+  lazy val amendRecycledPlasticPackagingRoute =
+    routes.AmendRecycledPlasticPackagingController.onPageLoad(NormalMode).url
 
-  lazy val amendHumanMedicinePlasticPackagingRoute =
-    routes.AmendHumanMedicinePlasticPackagingController.onPageLoad(NormalMode).url
-
-  "AmendHumanMedicinePlasticPackaging Controller" - {
+  "AmendRecycledPlasticPackaging Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
       val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
+        val request = FakeRequest(GET, amendRecycledPlasticPackagingRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
+        val view = application.injector.instanceOf[AmendRecycledPlasticPackagingView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, NormalMode, taxReturnOb)(request,
@@ -68,15 +67,14 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val ans = userAnswers.
-        set(AmendHumanMedicinePlasticPackagingPage, validAnswer).success.value
+      val ans = userAnswers.set(AmendRecycledPlasticPackagingPage, validAnswer).success.value
 
       val application = applicationBuilder(userAnswers = Some(ans)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
+        val request = FakeRequest(GET, amendRecycledPlasticPackagingRoute)
 
-        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
+        val view = application.injector.instanceOf[AmendRecycledPlasticPackagingView]
 
         val result = route(application, request).value
 
@@ -103,7 +101,7 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
 
       running(application) {
         val request =
-          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
+          FakeRequest(POST, amendRecycledPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
@@ -118,7 +116,7 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
       val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
+        val request = FakeRequest(GET, amendRecycledPlasticPackagingRoute)
 
         val result = route(application, request).value
 
@@ -132,12 +130,12 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
 
       running(application) {
         val request =
-          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
+          FakeRequest(POST, amendRecycledPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", "invalid value"))
 
         val boundForm = form.bind(Map("value" -> "invalid value"))
 
-        val view = application.injector.instanceOf[AmendHumanMedicinePlasticPackagingView]
+        val view = application.injector.instanceOf[AmendRecycledPlasticPackagingView]
 
         val result = route(application, request).value
 
@@ -153,12 +151,12 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
       val application = applicationBuilder(userAnswers = None).build()
 
       running(application) {
-        val request = FakeRequest(GET, amendHumanMedicinePlasticPackagingRoute)
+        val request = FakeRequest(GET, amendRecycledPlasticPackagingRoute)
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
 
@@ -168,14 +166,14 @@ class AmendHumanMedicinePlasticPackagingControllerSpec extends SpecBase with Moc
 
       running(application) {
         val request =
-          FakeRequest(POST, amendHumanMedicinePlasticPackagingRoute)
+          FakeRequest(POST, amendRecycledPlasticPackagingRoute)
             .withFormUrlEncodedBody(("value", validAnswer.toString))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
 
-        redirectLocation(result).value mustEqual routes.JourneyRecoveryController.onPageLoad().url
+        redirectLocation(result).value mustEqual controllers.routes.JourneyRecoveryController.onPageLoad().url
       }
     }
   }
